# docs/openapi.yml
openapi: "3.0.2"
info:
  title: Bjoetiek Y
  version: "1.0"
servers:
  - url: http://localhost:8090
  - url: https://backend.bjoetiek-y.be
paths:
...
  /public/manufacturers/{id}:
    get:
      description: Get manufacturer details
      tags: ["Manufacturers"]
      security: []
      parameters:
        - name: id
          in: path
          description: ID of the manufacturer
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Manufacturer"
  /admin/manufacturers:
    get:
      description: List manufacturers
      tags: ["Manufacturers"]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Manufacturer"

    post:
      description: Add manufacturer
      tags: ["Manufacturers"]
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Manufacturer"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Manufacturer"
  /admin/manufacturers/{id}:
    parameters:
      - name: id
        in: path
        description: ID of the manufacturer
        required: true
        schema:
          type: string
          format: uuid

    get:
      description: Get manufacturer details
      tags: ["Manufacturers"]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Manufacturer"

    put:
      description: Update manufacturer
      tags: ["Manufacturers"]
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Manufacturer"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Manufacturer"

    delete:
      description: Delete manufacturer
      tags: ["Manufacturers"]
      responses:
        "200":
          description: OK
...
components:
  schemas:
    Header:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "9ec47752-b691-4eb1-967a-99cdb3930de3"
          readOnly: True
...
    Manufacturer:
      allOf:
        - $ref: "#/components/schemas/Header"
        - type: object
          properties:
            name:
              type: string
              example: Bjoetiek Y
            website_url:
              type: string
              format: url
              example: https://bjoetiek-y.be
            logo_url:
              type: string
              format: url
              example: bjoetiek-y.jpg
...
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
security:
  - basicAuth: []